requirements.txt
# Nexus Finance MCP Server - Dependencies
# Install: pip install -r requirements.txt --break-system-packages

# === MCP Framework ===
fastmcp>=3.0.0,<4.0.0
mcp>=1.0.0,<2.0.0

# === Caching ===
diskcache>=5.6.0
cachetools>=5.3.0

# === Korean Public Data APIs ===
PublicDataReader>=1.0.0
OpenDartReader>=0.2.0

# === Stock Market Data ===
pykrx>=1.0.0
yfinance>=0.2.0

# === Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Statistics ===
scipy>=1.10.0
statsmodels>=0.14.0

# === Visualization ===
plotly>=5.18.0
matplotlib>=3.7.0
kaleido>=0.2.1

# === Global Macro Data ===
sdmx1>=2.0.0
wbgapi>=1.0.0

# === Crypto Exchange ===
ccxt>=4.0.0

# === Academic Search ===
scholarly>=1.7.0

# === News & Technical Analysis ===
feedparser>=6.0.0

# === Web/Async ===
requests>=2.31.0
aiohttp>=3.9.0

# === Environment ===
python-dotenv>=1.0.0
anyio>=4.0.0

# === Server (HTTP transport) ===
uvicorn>=0.30.0
starlette>=0.37.0
sse-starlette>=2.0.0
